[ISL-0108-1152] sistemati metodi create e delete

+++ Service/Impl/RegistroRicaricheService.cs
@@ -14,11 +14,13 @@ namespace MepWeb.Service.Impl
     {
         private readonly SataconsultingContext _dbContext;
         private readonly UserScope _scope;
+        private readonly IOreQualificaService _oreQualificaService;
 
-        public RegistroRicaricheService(SataconsultingContext dbContext, UserScope scope)
+        public RegistroRicaricheService(SataconsultingContext dbContext, UserScope scope, IOreQualificaService oreQualificaService)
         {
             _dbContext = dbContext;
             _scope = scope;
+            _oreQualificaService = oreQualificaService;
         }
 
         public async Task<ResponseBase<List<RegistroRicaricheResponse?>>> GetAllRecordsByIdDocAsync(decimal idDoc)
@@ -77,8 +79,14 @@ namespace MepWeb.Service.Impl
                 return ResponseBase<RegistroRicaricheResponse?>.Failed(GenericException.CampoObbligatorio, "Qualifica non passata, questo campo è obbligatorio");
             }
 
-            var checkExistence = await _dbContext.PscCo03s.FirstOrDefaultAsync(x => x.CDitta == cDitta && x.IdDoc == createRequest.IdDocumento && x.Grpcdl == createRequest.Qualifica);
-            if (checkExistence != null)
+            var c001 = await _oreQualificaService.GetSingleRecordAsync(createRequest.IdDocumento, createRequest.Qualifica);
+            if (!c001.Succeeded)
+            {
+                return ResponseBase<RegistroRicaricheResponse?>.Failed(GenericException.RecordInesistente, "Il record appartenente alla tabella PSC_C001 da cui deriva il record che si vuole creare non esiste");
+            }
+
+            var checkC003Existence = await _dbContext.PscCo03s.FirstOrDefaultAsync(x => x.CDitta == cDitta && x.IdDoc == createRequest.IdDocumento && x.Grpcdl == createRequest.Qualifica);
+            if (checkC003Existence != null)
             {
                 return ResponseBase<RegistroRicaricheResponse?>.Failed(GenericException.RecordGiaEsistente, "Il record che si sta cercando di creare è già presente");
             }
@@ -105,6 +113,18 @@ namespace MepWeb.Service.Impl
                 return ResponseBase<RegistroRicaricheResponse?>.Failed(GenericException.RecordNonCreato, "Non è stato possibile creare il record");
             }
 
+            OreQualificaUpdateRequest c001Update = new OreQualificaUpdateRequest();
+            c001Update.IdDocumento = createRequest.IdDocumento;
+            c001Update.Qualifica = createRequest.Qualifica;
+            c001Update.OreAcquistate = c001.Body.OreAcquistate + createRequest.OreAcquistate;
+
+            var update = await _oreQualificaService.UpdateRecordAsync(c001Update);
+            if (!update.Succeeded)
+            {
+                return ResponseBase<RegistroRicaricheResponse?>.Failed(GenericException.RecordNonAggiornato, "Non è stato possibile aggiornare il record nulla tabella di testata PSC_C001");
+            }
+
+
             return ResponseBase<RegistroRicaricheResponse?>.Success();
 
         }
@@ -112,6 +132,12 @@ namespace MepWeb.Service.Impl
         public async Task<ResponseBase<RegistroRicaricheResponse?>> UpdateRecordAsync(RegistroRicaricheUpdateRequest updateRequest)
         {
             string cDitta = CommonCostants.CDitta;
+
+            if (updateRequest.Id == 0)
+            {
+                return ResponseBase<RegistroRicaricheResponse?>.Failed(GenericException.CampoObbligatorio, "Id non passato, questo campo è obbligatorio");
+            }
+
             if (updateRequest.IdDocumento == 0)
             {
                 return ResponseBase<RegistroRicaricheResponse?>.Failed(GenericException.CampoObbligatorio, "Id Documento non passato, questo campo è obbligatorio");
@@ -122,7 +148,13 @@ namespace MepWeb.Service.Impl
                 return ResponseBase<RegistroRicaricheResponse?>.Failed(GenericException.CampoObbligatorio, "Qualifica non passata, questo campo è obbligatorio");
             }
 
-            var c003 = await _dbContext.PscCo03s.FirstOrDefaultAsync(x => x.CDitta == cDitta && x.IdDoc == updateRequest.IdDocumento && x.Grpcdl == updateRequest.Qualifica);
+            var checkC001Existance = await _oreQualificaService.GetSingleRecordAsync(updateRequest.IdDocumento, updateRequest.Qualifica);
+            if (checkC001Existance == null)
+            {
+                return ResponseBase<RegistroRicaricheResponse?>.Failed(GenericException.RecordInesistente, "Il record appartenente alla tabella PSC_C001 da cui deriva il record che si vuole aggiornare non esiste");
+            }
+
+            var c003 = await _dbContext.PscCo03s.FirstOrDefaultAsync(x => x.Id == updateRequest.Id);
             if (c003 == null)
             {
                 return ResponseBase<RegistroRicaricheResponse?>.Failed(GenericException.RecordInesistente, "Il record che si sta cercando di aggiornare non è presente all'interno della tabella");
@@ -176,6 +208,17 @@ namespace MepWeb.Service.Impl
                 return ResponseBase<RegistroRicaricheResponse?>.Failed(GenericException.RecordGiaEsistente, "Il record che si sta cercando di eliminare non è presente all'interno della tabella");
             }
 
+            var c001 = await _oreQualificaService.GetSingleRecordAsync(c003.IdDoc, c003.Grpcdl);
+            if (!c001.Succeeded)
+            {
+                return ResponseBase<RegistroRicaricheResponse?>.Failed(GenericException.RecordInesistente, "Il record appartenente alla tabella PSC_C001 da cui deriva il record che si vuole eliminare non esiste");
+            }
+
+            OreQualificaUpdateRequest c001Update = new OreQualificaUpdateRequest();
+            c001Update.IdDocumento = c003.IdDoc;
+            c001Update.Qualifica = c003.Grpcdl;
+            c001Update.OreAcquistate = c001.Body.OreAcquistate - c003.HhAcq;
+
             try
             {
                 _dbContext.PscCo03s.Remove(c003);
@@ -186,17 +229,11 @@ namespace MepWeb.Service.Impl
                 return ResponseBase<RegistroRicaricheResponse?>.Failed(GenericException.RecordNonCreato, "Non è stato possibile eliminare il record su PSC_C003");
             }
 
-            //var c001 = await _dbContext.PscCo01s.FirstOrDefaultAsync(x => x.CDitta == cDitta && x.IdDoc == idDoc && x.Grpcdl == grpcdl);
-
-            //try
-            //{
-            //    _dbContext.PscCo01s.Remove(c001);
-            //    var res = await _dbContext.SaveChangesAsync();
-            //}
-            //catch (Exception ex)
-            //{
-            //    return ResponseBase<RegistroRicaricheResponse?>.Failed(GenericException.RecordNonEliminato, "Non è stato possibile eliminare il record su PSC_C001");
-            //}
+            var update = await _oreQualificaService.UpdateRecordAsync(c001Update);
+            if (!update.Succeeded)
+            {
+                return ResponseBase<RegistroRicaricheResponse?>.Failed(GenericException.RecordNonAggiornato, "Non è stato possibile aggiornare il record nulla tabella di testata PSC_C001");
+            }
 
             return ResponseBase<RegistroRicaricheResponse?>.Success();
         }
